Add redirect_helper file
